Login

+++ src/main/java/com/proyecto/transporte/dao/BusRepository.java
@@ -3,6 +3,6 @@ package com.proyecto.transporte.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyecto.transporte.entidad.Bus;
 
-public interface BusRepository extends JpaRepository<Bus, Long>{
+public interface BusRepository extends JpaRepository<Bus, Integer>{
 
 }
