+++ src/main/java/com/proyecto/transporte/dao/DestinoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.proyecto.transporte.entidad.Destino;
 
-public interface DestinoRepository extends JpaRepository<Destino, Long>{
+public interface DestinoRepository extends JpaRepository<Destino, Integer>{
 
 }
